envoi le score via signal R au joueur

+++ GamingPlatform/Migrations/GamingPlatformContextModelSnapshot.cs
@@ -192,8 +192,8 @@ namespace GamingPlatform.Migrations
                         .IsRequired()
                         .HasColumnType("nvarchar(max)");
 
-                    b.Property<int>("Score")
-                        .HasColumnType("int");
+                    b.Property<double>("Score")
+                        .HasColumnType("float");
 
                     b.Property<string>("SessionToken")
                         .IsRequired()
